decorator inside class
