+++ fake_attendance/helper.py
@@ -11,18 +11,6 @@ import pyautogui
 from win32com.client import Dispatch
 import win32gui
 
-def decorator_start_end(name):
-    'outer decorator, argument should be name for script'
-    def inner_decorator(func):
-        'inner'
-        def wrapper(*args, **kwargs):
-            'wrapper'
-            print_with_time(f'{name} 스크립트 시작')
-            func(*args, **kwargs)
-            print_with_time(f'{name} 스크립트 종료')
-        return wrapper
-    return inner_decorator
-
 def get_file_path(filename, sub=None):
     'return full file path'
     if sub:
